refactor: eventlistener 위치조정
